requirements.txt
discord.py[voice]==1.3.2
django
colorlog
chardet
cffi
aiohttp
websockets
pymongo
boto
heroku3
dnspython
pynacl
opuslib
twitch-python
zc.lockfile